requirements.txt
Flask
Flask-PyMongo
pymongo
flask-login
flask-bcrypt
python-dotenv
